requirements.txt
azure-search-documents==11.5.1
semantic-kernel==1.8.3
termcolor==2.4.0